[Fix][Android] Padding and border should occupy position as View padding

+++ Android/sdk/src/main/java/com/lynx/base/Style.java
@@ -36,7 +36,7 @@ public class Style {
 
 
     public int mBackgroundColor;
-    public double mBorderWidth;
+    public int mBorderWidth;
     public int mBorderColor;
     public double mBorderRadius;
     public double mOpacity ;
